refactor: extract postCreatePR to pr-post-create.ts
Move the post-PR-creation side-effects (copy_milestone, copy_assignees,
copy_all_reviewers, copy_requested_reviewers, labels, author-as-assignee,
author-as-reviewer, add_reviewers, add_team_reviewers, auto_merge_enabled)
out of Backport.run() into a standalone postCreatePR function.

Each side-effect is best-effort: a RequestError is logged and the
remaining steps continue. labelsToCopy is filtered in run() and passed
in pre-resolved. workflowRepo is passed for listReviews on the source
PR (used by copy_all_reviewers) since reviews live on the workflow
repo, not the target.

The auto-merge error-mapping helper (getAutoMergeErrorMessage) moves
along as a private function since it is only used here.

No user-visible behavioral change.
